refactor: 컴포넌트이름 변경 (PostButtons => MyPageBottomUIs)
